refactor(theme): extract CSS generation into generateThemeRules.ts
Move theme CSS generation logic (generateThemeRules, generateThemeRulesSplit,
generateOnMediaCSS, generateThemeCSS, generateThemeCSSFlat) and their helpers
(parsePadding, expandContainerPadding, toKebabCase, PADDING_PROPS) from
defineTheme.ts into a dedicated generateThemeRules.ts module.

This reduces defineTheme.ts from 1235 to 668 lines and lowers its cyclomatic
complexity by isolating CSS generation concerns. The module decomposition also
breaks the monolithic generateThemeRules function into focused sub-functions:
generateComponentRules, generateProseRules, and generateColorOverrides.

The CLI's legacy fallback implementations (used only when @xds/core isn't
built) are marked with deprecation comments pointing to the canonical source.

No behavior changes — pure structural refactoring. All 223 theme tests pass.
